+++ src/sumukha/preprocessing/input_preprocessor.py
@@ -12,6 +12,8 @@ from sklearn.pipeline import Pipeline
 
 from sumukha.config import PUNCTUATION, input_preprocess_path
 
+import os
+
 import re
 
 
@@ -51,7 +53,19 @@ def clean_data(data):
 
     data['review'] = features_encoder.transform(reviews.copy())
 
-    data['review'].to_csv(input_preprocess_path + 'data.txt', index=False, header=False)
+    with open(input_preprocess_path + 'data.txt', 'w') as f:
+        for index, row in data.iterrows():
+            f.write('__label__' + str(row['sentiment']) + ' ' + str(row['review']) + '\n')
+
+    f.close()
+
+    training_size = int(data.shape[0] * 0.80)
+    test_size = int(data.shape[0] * 0.20)
+    file_path = input_preprocess_path + 'data.txt'
+    training_path = input_preprocess_path + 'data_train.txt'
+    test_path = input_preprocess_path + 'data_test.txt'
+    os.system(f'head -n {training_size} {file_path} > {training_path}')
+    os.system(f'tail -n {test_size} {file_path}  > {test_path}')
 
     return data